# openapi.yaml
swagger: '2.0'
info:
  title: CRM Gateway Config
  description: API Gateway mapping for CRM API
  version: 1.0.0
schemes:
  - https
paths:
...
  /api/products/{id}:
    get:
      operationId: get_id_api_products
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/products/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_products
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/products/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_products
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/products/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/catalog/products/{id}:
    get:
      operationId: get_id_api_v1_catalog_products
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/products/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_catalog_products
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/products/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_catalog_products
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/products/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/crm/products/{id}:
    get:
      operationId: get_id_api_v1_crm_products
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/products/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_crm_products
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/products/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_crm_products
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/products/{id}
      responses: { '200': { description: 'OK' } }
...